Commented the extra doctype on blog_post and blog
Not sure why they specifically had their own doctype specified, but the files also imported the HTMLHead panel which has its own doctype in that. So basically it had 2 doctypes. I noticed this because blog and blog_post header was just 3px shorter. Not sure why, but at least now it’s consistent